+++ CHANGELOG.md
@@ -4,7 +4,6 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### 0.1.2 (2026-02-06)
 
-
 ### Bug Fixes
 
-* attempt to fix numerical suffixes in generated types ([d2a555f](https://github.com/safwanyp/json-schema-to-dts/commit/d2a555f1e64997de63df090ee1fb0bb180b1199a))
+- attempt to fix numerical suffixes in generated types ([d2a555f](https://github.com/safwanyp/json-schema-to-dts/commit/d2a555f1e64997de63df090ee1fb0bb180b1199a))